requirements.txt
aiofiles==23.2.1
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bidict==0.23.1
certifi==2024.07.04
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
isodate==0.6.1
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.5
markdown2==2.4.12
MarkupSafe==2.1.5
msrest==0.7.1
multidict==6.0.5
nicegui>=1.3
oauthlib==3.2.2
orjson==3.10.3
packaging==23.2
paho-mqtt>=1.6
pscript==0.7.7
pydantic>=2.7.1,<3.0.0
Pygments==2.17.2
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.18
python-socketio==5.11.1
PyYAML==6.0.1
requests
requests-oauthlib==1.3.1
requests-unixsocket
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy>=2.0
starlette==0.36.3
typing_extensions==4.11.0
urllib3
uvicorn==0.27.1
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
XlsxWriter==3.2.0
yarl==1.9.5
loguru>=0.7
psycopg2-binary>=2.9.5
